Adding fix for AutoMode not publishing
The problem was in the `std_msgs/Header`. If that is included, we aren't able to see the topic being published.

According to the documentation, we can omit the header when publishing altogether and `roslibpy` will automatically populate the fields.
Reference: https://github.com/RobotWebTools/rosbridge_suite/blob/f0d359797eb90b322e5eb23cfda89ea7a9aceac8/ROSBRIDGE_PROTOCOL.md#343-publish--publish-

After removing the `header` field when publishing, we are now able to see the `auto_mode`. For some reason, the time stamp is not being updated, but we are now able to publish with headers.